fix allowed hosts
