+++ website/app/settings/base.py
@@ -13,6 +13,8 @@ https://docs.djangoproject.com/en/5.1/ref/settings/
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 import os
 import dj_database_url
+import json
+import urllib.request
 
 PROJECT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 BASE_DIR = os.path.dirname(PROJECT_DIR)
@@ -291,6 +293,27 @@ GITHUB_SHA = os.getenv("GITHUB_SHA")
 EMAIL_BACKEND = "django.core.mail.backends.console.EmailBackend"
 
 ALLOWED_HOSTS = ["localhost", "127.0.0.1", os.getenv("DOMAIN_NAME")]
+
+
+def _task_ips():
+    """Return the task’s IPv4 address(es) from the ECS metadata API."""
+    meta = os.getenv("ECS_CONTAINER_METADATA_URI_V4") or os.getenv(
+        "ECS_CONTAINER_METADATA_URI"
+    )
+    if not meta:
+        return []
+
+    try:
+        with urllib.request.urlopen(f"{meta}/task", timeout=0.2) as r:
+            data = json.load(r)
+            # First container in the task is usually “ours”
+            nets = data["Containers"][0]["Networks"]
+            return [ip for net in nets for ip in net["IPv4Addresses"]]
+    except Exception:
+        return []
+
+
+ALLOWED_HOSTS += _task_ips()
 
 USE_X_FORWARDED_HOST = True
 SECRET_KEY = os.getenv("SECRET_KEY")
